+++ Modules/Ruleset.ProgramInfo/Ruleset.ProgramInfo.psm1
@@ -35,6 +35,10 @@ Set-Variable -Name ThisModule -Scope Script -Option ReadOnly -Force -Value ($MyI
 
 #
 # Script imports
+# TODO: Executables involved in rules which are install into ProgramFiles\Common Files require
+# separate search algorithm (function) instead of using Update-Table, these are present only
+# if program in question is installed, likely into ProgramFiles
+# Example programs: Adobe and Java
 #
 
 $PrivateScripts = @(