+++ Rules/IPv4/Outbound/Software/Adobe.ps1
@@ -71,7 +71,8 @@ if (!(Approve-Execute -Accept $Accept -Deny $Deny)) { exit }
 # Adobe installation directories
 #
 $AdobeARMRoot = "%ProgramFiles(x86)%\Common Files\Adobe\ARM\1.0"
-$AcrobatRoot = "%ProgramFiles(x86)%\Adobe\Acrobat Reader DC\Reader"
+$AcrobatRoot = "%ProgramFiles(x86)%\Adobe\Acrobat DC"
+$ReaderRoot = "%ProgramFiles(x86)%\Adobe\Acrobat Reader DC"
 
 # First remove all existing rules matching group
 Remove-NetFirewallRule -PolicyStore $PolicyStore -Group $Group -Direction $Direction -ErrorAction Ignore
@@ -81,29 +82,104 @@ Remove-NetFirewallRule -PolicyStore $PolicyStore -Group $Group -Direction $Direc
 #
 
 # Test if installation exists on system
+if ((Test-Installation "AdobeReader" ([ref] $ReaderRoot)) -or $ForceLoad)
+{
+	$Program = "$ReaderRoot\Reader\AcroRd32.exe"
+	Test-File $Program
+
+	New-NetFirewallRule -DisplayName "Adobe Reader" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled True -Action Allow -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 80, 443 `
+		-LocalUser $UsersGroupSDDL `
+		-InterfaceType $DefaultInterface `
+		-Description "" | Format-Output
+
+	$Program = "$ReaderRoot\Reader\AcroCEF\RdrCEF.exe"
+	Test-File $Program
+
+	New-NetFirewallRule -DisplayName "Adobe Reader cloud services" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled False -Action Allow -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 443 `
+		-LocalUser $UsersGroupSDDL `
+		-InterfaceType $DefaultInterface `
+		-Description "RdrCEF.exe is inseparable processes of Adobe Reader.
+It handles multiple integral aspects of application like network interaction and
+Document Cloud services like Fill and Sign, Send For Signature, Share for view/review, and so on)" |
+	Format-Output
+}
+
+# Test if installation exists on system
 if ((Test-Installation "AdobeAcrobat" ([ref] $AcrobatRoot)) -or $ForceLoad)
 {
-	$Program = "$AcrobatRoot\AcroRd32.exe"
+	$Program = "$AcrobatRoot\Acrobat\Acrobat.exe"
 	Test-File $Program
-	New-NetFirewallRule -Platform $Platform `
-		-DisplayName "Acrobat Reader" -Service Any -Program $Program `
-		-PolicyStore $PolicyStore -Enabled False -Action Allow -Group $Group -Profile $DefaultProfile -InterfaceType $DefaultInterface `
-		-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 80, 443 `
+
+	New-NetFirewallRule -DisplayName "Adobe Acrobat Pro" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled True -Action Allow -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 80, 443 `
 		-LocalUser $UsersGroupSDDL `
+		-InterfaceType $DefaultInterface `
 		-Description "" | Format-Output
+
+	$Program = "$AcrobatRoot\Acrobat\AcroCEF\AcroCEF.exe"
+	Test-File $Program
+
+	New-NetFirewallRule -DisplayName "Adobe Acrobat Pro cloud services" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled False -Action Allow -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 443 `
+		-LocalUser $UsersGroupSDDL `
+		-InterfaceType $DefaultInterface `
+		-Description "AcroCEF.exe is inseparable processes of Acrobat Reader.
+It handles multiple integral aspects of application like network interaction and
+Document Cloud services like Fill and Sign, Send For Signature, Share for view/review, and so on)" |
+	Format-Output
+
+	# TODO: This is workaround, see todo comment in Programinfo.psm1, separate search needed
+	$Program = "%SystemDrive%\Program Files (x86)\Common Files\Adobe\AdobeGCClient\AdobeGCClient.exe"
+	Test-File $Program
+
+	New-NetFirewallRule -DisplayName "Adobe Genuine Software Integrity" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled True -Action Allow -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 443 `
+		-LocalUser $UsersGroupSDDL `
+		-InterfaceType $DefaultInterface `
+		-Description "Adobe Genuine Software Integrity Service" |
+	Format-Output
 }
 
 # Test if installation exists on system
 if ((Test-Installation "AdobeARM" ([ref] $AdobeARMRoot)) -or $ForceLoad)
 {
-	$Program = "$AdobeARMRoot\AdobeARM.exe.exe"
+	$Program = "$AdobeARMRoot\AdobeARM.exe"
 	Test-File $Program
-	New-NetFirewallRule -Platform $Platform `
-		-DisplayName "Acrobat ARM" -Service Any -Program $Program `
-		-PolicyStore $PolicyStore -Enabled False -Action Allow -Group $Group -Profile $DefaultProfile -InterfaceType $DefaultInterface `
-		-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 80 `
+
+	New-NetFirewallRule -DisplayName "Acrobat ARM" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled True -Action Allow -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 80 `
 		-LocalUser $UsersGroupSDDL `
-		-Description "" | Format-Output
+		-InterfaceType $DefaultInterface `
+		-Description "Adobe updater is responsible for checking for, downloading, and launching the
+update installer for Reader or Acrobat.
+The Updater primarily keeps itself up to date and downloads and extracts needed files.
+It does not actually install anything, as that job is handled by a separate installer" | Format-Output
 }
 
 Update-Log